refactor: вынесение хранения данных о пользователь в отдельный класс
